+++ build_my_garden/lib/service/seller_info_service.dart
@@ -69,7 +69,7 @@ class SellerInfo {
   String first_name;
   String last_name;
   String city;
-  int seller_rating;
+  double seller_rating;
   int user;
 
   SellerInfo({
@@ -90,7 +90,7 @@ class SellerInfo {
       last_name: map['last_name'],
       city: map['city'],
       user: map['user'],
-      seller_rating: map['seller_rating'],
+      seller_rating: map['seller_rating'].toDouble(),
     );
   }
 }
